feat: implements useWreathSans hook
