+++ projects/waffle-sans/src/hooks/useWreathSans.ts
@@ -1,0 +1,37 @@
+import { useEffect, useMemo, useRef } from 'react';
+
+import createWreathSans from '../../../../libs/leon-sans-react/src/hooks/createWreathSans';
+
+interface Params {
+  initialText: string;
+}
+
+export default function useWreathSans({ initialText }: Params) {
+  const ref = useRef<HTMLDivElement>(null);
+
+  const { WreathSansCanvas, resize, redraw } = useMemo(() => {
+    return createWreathSans({
+      initialText: initialText,
+      width: ref?.current?.offsetWidth,
+      height: ref?.current?.offsetHeight,
+      size: 40,
+      color: '#704234',
+      background: 'transparent',
+    });
+  }, [initialText]);
+
+  useEffect(() => {
+    function handleResize() {
+      if (ref.current?.offsetWidth && ref.current?.offsetHeight) {
+        resize(ref.current.offsetWidth, ref.current.offsetHeight);
+        redraw();
+      }
+    }
+    window.addEventListener('resize', handleResize);
+    return () => {
+      window.removeEventListener('resize', handleResize);
+    };
+  }, [redraw, resize]);
+
+  return { WreathSansCanvas, ref, resize, redraw };
+}
